feat: add optional topology filter
Adds a `filter` option to topologies to allow filtering out duplicate
notifications.
